Time: 0 ms (100.00%) | Memory: 7.8 MB (68.55%) - LeetSync

+++ 3918-check-divisibility-by-digit-sum-and-product/check-divisibility-by-digit-sum-and-product.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool checkDivisibility(int n) {
-        long long prod = 1, sum = 0;
+        int prod = 1, sum = 0;
         int m = n;
         while(m > 0) {
             sum += m % 10;
